chore(wallaby.js): fix test issues
